refactor: review-comment: Filename to match main class name
