+++ src/opossum_lib/input_formats/scancode/entities/scan_code_data_raw.py
@@ -155,7 +155,7 @@ class File(BaseModel):
     urls: list[Url]
 
 
-class ScanCodeData(BaseModel):
+class ScanCodeDataRaw(BaseModel):
     dependencies: list
     files: list[File]
     license_detections: list[GlobalLicenseDetection]
