+++ src/opossum_lib/shared/entities/opossum_information.py
