+++ src/opossum_lib/shared/entities/opossum_output_file.py
