+++ tests/opossum_model_generators/scan_results_provider.py
@@ -8,7 +8,7 @@ from faker.providers import BaseProvider
 from faker.providers.file.en_US import Provider as FileProvider
 from faker.providers.misc.en_US import Provider as MiscProvider
 
-from opossum_lib.core.entities.opossum_model import (
+from opossum_lib.core.entities.opossum import (
     BaseUrlsForSources,
     ExternalAttributionSource,
     FrequentLicense,
