+++ tests/test_opossum/generators/generate_opossum_file_content.py
@@ -5,9 +5,9 @@ from typing import Any
 
 from faker.providers import BaseProvider
 
-from opossum_lib.shared.entities.opossum_file import OpossumInformation
 from opossum_lib.shared.entities.opossum_file_content import OpossumFileContent
-from opossum_lib.shared.entities.output_model import OpossumOutputFile
+from opossum_lib.shared.entities.opossum_information import OpossumInformation
+from opossum_lib.shared.entities.opossum_output_file import OpossumOutputFile
 from tests.test_opossum.generators.generate_file_information import (
     FileInformationProvider,
 )
