+++ tests/test_opossum/generators/generate_outfile_information.py
@@ -9,7 +9,7 @@ from faker.providers.date_time import Provider as DatetimeProvider
 from faker.providers.lorem.en_US import Provider as LoremProvider
 from faker.providers.misc import Provider as MiscProvider
 
-from opossum_lib.shared.entities.output_model import (
+from opossum_lib.shared.entities.opossum_output_file import (
     ManualAttributions,
     Metadata,
     OpossumOutputFile,
