+++ tests/test_scancode/test_get_attribution_info.py
@@ -7,7 +7,7 @@ from opossum_lib.input_formats.scancode.constants import SCANCODE_SOURCE_NAME
 from opossum_lib.input_formats.scancode.services.scancode_data_to_opossum_converter import (  # noqa: E501
     ScancodeDataToOpossumConverter,
 )
-from opossum_lib.shared.entities.opossum_file import OpossumPackage, SourceInfo
+from opossum_lib.shared.entities.opossum_information import OpossumPackage, SourceInfo
 from tests.test_setup.scancode_faker_setup import ScanCodeFaker
